Move functions to util module
